chore: add RAMP type to global windowwq
