+++ .vscode/settings.json
@@ -4,6 +4,7 @@
         "Geosearch",
         "Proxied",
         "TABBABLE",
+        "WKID",
         "activedescendant",
         "animejs",
         "basemap",
@@ -21,6 +22,7 @@
         "pathify",
         "prefetch",
         "purgecss",
+        "rampgeo",
         "tailwindcss",
         "tslib",
         "unminified",
